ci: fetch-depth 0 to fetch complete history
With `${{ github.event.pull_request.commits }}` we don't fetch enough
commits to diff against the merge base. We require `pull_request.commits + 1`
commits to be able do that.

An alternative is to just fetch the complete git history as we don't
expect this repository to grow to a point where this has a
significant impact on CI performance.